Lots of items cleaned up - Good progress!
With help from John Jardin and Heiko Voigt, got lots of items working.  The List page is broken, which happened after I changed to John's idea of a memoryStore to hold the token.  I think React-Router is breaking things now, as when I manually navigate to that page all of state and the memoryStore is wiped out.